+++ src/base/operation.js
@@ -41,6 +41,7 @@ import { CreatePaymentRequestBuilder } from './operations/create_payment_request
 import { RemoveAssetOpBuilder } from './operations/remove_asset_op_builder'
 import { OpenSwapBuilder } from './operations/open_swap_builder'
 import { CloseSwapBuilder } from './operations/close_swap_builder'
+import { RedemptionRequestBuilder } from './operations/redemption_request_op_builder'
 
 export class Operation extends BaseOperation {
   /**
@@ -494,6 +495,9 @@ export class Operation extends BaseOperation {
         break
       case xdr.OperationType.closeSwap():
         CloseSwapBuilder.closeSwapToObject(result, attrs)
+        break
+      case xdr.OperationType.createRedemptionRequest():
+        RedemptionRequestBuilder.redemptionRequestToObject(result, attrs)
         break
       default:
         throw new Error('Unknown operation')