+++ CHANGELOG.md
@@ -6,6 +6,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 Please check our [developers guide](https://gitlab.com/tokend/developers-guide)
 for further information about branching and tagging conventions.
+
+## [Unreleased]
+### Added
+- Redemption reviewable request
+
 ## [1.11.3] - 2019-12-04
 ### Added
 - `signAndSendTransaction` method in to api-caller